[WFCORE-1618] Move ManagementRequestContextImpl out of AbstractMessageHandler to make AMH less complex
